# Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UnnamedMessenger API",
    "description": "API specification for secure messenger with E2E encryption",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/messages/{messageId}": {
      "put": {
        "summary": "Edit a message",
        "tags": ["Messages"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "New text content (max 10000 chars)"
                  },
                  "fileIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of file IDs to attach (replace existing). Every file must exist and belong to the same chat."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message edited successfully"
          },
          "400": {
            "description": "Invalid input, message limit exceeded, or some fileIds do not exist / do not belong to this chat"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not allowed to edit this message"
          },
          "404": {
            "description": "Message not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a message",
        "tags": ["Messages"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not allowed to delete this message"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
...
    "/files/me": {
      "get": {
        "summary": "Get current user's files and quota",
        "tags": ["Files"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user's files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quotaBytes": {
                      "type": "integer"
                    },
                    "usedBytes": {
                      "type": "integer"
                    },
                    "files": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/FileRecord"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "isAvatar": {
                                "type": "boolean"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/files/me/{fileId}": {
      "delete": {
        "summary": "Delete own file completely",
        "tags": ["Files"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "fileId": {
                      "type": "string"
                    },
                    "quotaBytes": {
                      "type": "integer"
                    },
                    "usedBytes": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/files/me/{fileId}/placeholder": {
      "put": {
        "summary": "Replace own file with placeholder",
        "tags": ["Files"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-file-metadata",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base64-encoded encrypted placeholder metadata"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File replaced with placeholder",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file": {
                      "$ref": "#/components/schemas/FileRecord"
                    },
                    "quotaBytes": {
                      "type": "integer"
                    },
                    "usedBytes": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid placeholder content or metadata"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FileRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chatId": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          },
          "deletedAt": {
            "type": "integer",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}